+++ app/src/main/java/com/example/alexander/weatherapp/data/network/models/weather/Wind.java
@@ -13,13 +13,13 @@ public class Wind {
 
     @SerializedName("deg")
     @Expose
-    private int deg;
+    private double deg;
 
     public double getSpeed() {
         return speed;
     }
 
-    public int getDeg() {
+    public double getDeg() {
         return deg;
     }
 }
